+++ home/templates/home/tags/section_progress.html
@@ -1,7 +1,7 @@
-<div class="user-progress-container">
-    {% if total %}
+{% if total %}
+    <div class="user-progress-container">
         <div>
             User Progress: {{ read }}/{{ total }} completed
         </div>
-    {% endif %}
-</div>
+    </div>
+{% endif %}
