+++ iogt_users/models.py
@@ -13,7 +13,24 @@ class User(AbstractUser):
 
     has_filled_registration_survey = models.BooleanField(default=False)
 
-    viewed_articles = models.ManyToManyField(to='home.Article')
+    read_articles = models.ManyToManyField(to='home.Article')
+
+    @classmethod
+    def record_article_read(cls, request, article):
+        user = request.user
+        if user.is_anonymous:
+            read_articles = request.session.get('read_articles', [])
+            if read_articles:
+                if article.pk not in read_articles:
+                    # https://code.djangoproject.com/wiki/NewbieMistakes#Appendingtoalistinsessiondoesntwork
+                    read_articles = request.session['read_articles']
+                    read_articles.append(article.pk)
+                    request.session['read_articles'] = read_articles
+            else:
+                request.session['read_articles'] = [article.pk]
+        else:
+            if not user.read_articles.filter(id=article.id).exists():
+                user.read_articles.add(article)
 
     class Meta:
         ordering = ('id',)
